16 bit integer for cursorpos pointer

+++ schedos-1.c
@@ -26,7 +26,7 @@ start(void)
 	for (i = 0; i < RUNCOUNT; i++) {
 		// Write characters to the console, yielding after each one.
 		//*cursorpos++ = PRINTCHAR;
-		sys_print(PRINTCHAR);
+		sys_print((uint16_t)PRINTCHAR);
 		sys_yield();
 	}
